Drop the no-op casts the `bytes_at` sweep left behind
`metrics/clippy.json` had grown 19 warnings / 16 files to 48 / 39 across
p24-6 and p24-7, and the 29 new ones were all mechanical residue of the
`STRLEN` -> `cstr::bytes_at(..).len()` conversion:

* **26 `unnecessary_cast`.** `STRLEN` answered `size_t`, so every call
  site wore an `as usize`; `.len()` already *is* `usize` and the cast
  became `usize -> usize`. Four sites are the same shape one step out
  (`add(len as usize)`, `positions[pat_len as usize - 1]`,
  `str.offset(len as isize)`, `src_len as usize + 1`) where the bound
  variable had already moved to the target type.
* **2 `len_zero`.** `STRLEN(p) != 0` transliterates to
  `bytes_at(p).len() != 0`; on a slice that is `!is_empty()`.
* **1 `empty_line_after_doc_comments`.** `api/vim/inspect.rs`'s
  `NULL_STR` const went away with the `_c` message macros and left its
  doc comment behind, dangling above `nvim__id`.

None was correct-as-written, so nothing needed an `#[allow]`. The
baseline is back to **19 / 16**, exactly the phase base's, and the
`--allow-growth` that carried the drift is retired.

rustfmt re-joined four calls that fit on one line once the casts were
gone, which takes `unsafe_lines` 68,718 -> 68,717 and
`unsafe_lines_outside_perimeter` 55,423 -> 55,422.

+++ crates/nvim/src/arglist/mod.rs
@@ -429,7 +429,7 @@ unsafe fn get_arglist(gap: *mut garray_T, str: *mut c_char, escaped: bool) {
     // argument: `split_one_arg` only ever shortens what it is given, so the
     // original length still bounds it.
     unsafe { ga_init(gap, size_of::<*mut c_char>() as c_int, 20) };
-    let total = unsafe { cstr::bytes_at(str) }.len() as usize;
+    let total = unsafe { cstr::bytes_at(str) }.len();
     let buf = unsafe { core::slice::from_raw_parts_mut(str.cast::<u8>(), total + 1) };
     let mut at = 0;
     while at < total && buf[at] != 0 {
